requirements.txt
blinker==1.7.0
blurhash==1.1.4
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
filetype==1.2.0
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Mastodon.py==1.8.1
Pillow==10.1.0
python-dateutil==2.8.2
python-magic-bin==0.4.14
requests==2.31.0
six==1.16.0
urllib3==2.1.0
Werkzeug==3.0.1
